+++ packages/adminui-vue/components/UMasterDetailView/UMasterDetailView.vue
@@ -287,8 +287,10 @@ export default {
     },
 
     buildDetailAddNewConfig (cfg) {
-      cfg.parentContext = {
-        [this.selectedDetail.attribute]: this.selectedRowId
+      cfg.props = {
+        parentContext: {
+          [this.selectedDetail.attribute]: this.selectedRowId
+        }
       }
       return cfg
     },